Refactor code into transcript.py
